style(perf): split inline test modules into path-based sibling files (#300)
This repo's change-risk gate (change_risk/signals.rs::is_test_file)
only recognizes a companion test file by path convention -- tests/**,
test_*/*_test(s).*, or a bare tests.rs -- not an inline
#[cfg(test)] mod tests {} block in the same file. Both files this
ticket touched had their tests inline, so the gate scored "source
changed, no tests touched" (its largest single weight) and blocked
the PR at 80/100 against the 80 threshold.

- doctor_provider_rows.rs -> doctor_provider_rows_tests.rs (matches
  the existing artifacts.rs -> artifacts_tests.rs convention)
- doctor_bootstrap/mod.rs -> doctor_bootstrap/tests.rs (matches the
  existing change_risk/tests.rs, change_agenda/tests.rs convention
  for a module that already lives in its own directory)

No behavior change: same tests, moved verbatim, wired in via
#[path]. Verified across every #[path]-inclusion context this repo
mounts doctor_bootstrap/doctor_provider_rows into (doctor_bootstrap::
tests, capability_inventory::doctor_adapter::doctor_bootstrap::tests)
via the full cargo test --bins --tests sweep, not just --bin
code-intel.

Bonus: doctor_bootstrap/mod.rs drops from 963 to 744 lines, clearing
this repo's own god-file threshold (33 -> 32 god files) as a side
effect -- not the point of this commit, but sentrux check reports it.

Refs #299 #300.
